backend/requirements.txt
# ConceptCanvas Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# OpenAI & AI
openai>=1.59.0

# Vector Database
pinecone-client>=6.0.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0
tenacity>=9.0.0
sse-starlette>=2.2.0

# Database (NEW)
sqlalchemy>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.20.0
alembic>=1.14.0

# Authentication (NEW)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Email (NEW)
aiosmtplib>=3.0.2

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
